requirements.txt
Pillow
numpy<2.0.0
matplotlib
timm>=0.9,<=1.0.12
wheel
cython
tensorboard
scipy
opencv-python
scikit-learn
scikit-image
openai-clip
decord
requests
pyyaml
easydict
pyiqa
lvis
fairscale>=0.4.4
fvcore
easydict
urllib3
boto3
omegaconf
transformers==4.33.2
pycocoevalcap
openpyxl
